remove unused evaluate,model,train pys, refactor tf utilities
